Minor refactoring and PR feedback
