+++ projects/Mallard/src/screens/editions-menu-screen.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react'
 import { StyleSheet, View } from 'react-native'
 import { NavigationScreenProp } from 'react-navigation'
 import { EditionsMenu } from 'src/components/EditionsMenu/EditionsMenu'
-import { EditionsMenuScreenHeader } from 'src/components/layout/header/EditionMenuScreenHeader'
+import { EditionsMenuScreenHeader } from 'src/components/ScreenHeader/EditionMenuScreenHeader'
 import { useEditions } from 'src/hooks/use-edition-provider'
 import { routeNames } from 'src/navigation/routes'
 import { WithAppAppearance } from 'src/theme/appearance'
